abstracted out the RawTextDataset class
